requirements.txt
blinker==1.8.2
click==8.1.8
colorama==0.4.6
Flask==3.0.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.2
Werkzeug==3.0.6
xgboost==2.1.4
zipp==3.20.2
